+++ src/http_types.zig
@@ -5,6 +5,10 @@ pub const Request = struct {
     target: []const u8,
     content_type: ?[]const u8 = null,
     body: []const u8 = "",
+    host: ?[]const u8 = null,
+    date: ?[]const u8 = null,
+    digest: ?[]const u8 = null,
+    signature: ?[]const u8 = null,
     cookie: ?[]const u8 = null,
     authorization: ?[]const u8 = null,
 };